Extract individual block, element and composition object classes into their own files as there are too many of them to sustainably maintain them in the single `Block.kt`, `Element.kt` and `CompositionObject.kt` files Refs #4, Refs #5, Refs #7
