+++ src/main/kotlin/eth/rochsolid/slack/blocks/models/blocks/Section.kt
@@ -1,0 +1,106 @@
+package eth.rochsolid.slack.blocks.models.blocks
+
+import eth.rochsolid.slack.blocks.models.elements.Element
+import kotlinx.serialization.SerialName
+
+
+/**
+ * Displays text, possibly alongside block elements.
+ * A section can be used as a text block, in combination with text fields, or side-by-side with certain [Element]s.
+ */
+sealed class Section(
+    /**
+     * A unique identifier for a block. If not specified, a block_id will be generated by the Slack API.
+     * You can use this [blockID] when you receive an interaction payload to identify the source of the action.
+     * Maximum length for this field is 255 characters. [blockID] should be unique for each message and each iteration
+     * of a message. If a message is updated, use a new [blockID].
+     */
+    @SerialName("block_id")
+    override val blockID: BlockID?
+) : Block(type = Type.SECTION) {
+    /**
+     * Example:
+     * ```json
+     * {
+     *   "type": "section",
+     *   "text": {
+     *     "type": "mrkdwn",
+     *     "text": "A message *with some bold text* and _some italicized text_."
+     *   }
+     * }
+     * [View this block in Block Kit Builder](https://api.slack.com/tools/block-kit-builder?blocks=%5B%0A%09%7B%0A%09%09%22type%22%3A%20%22section%22%2C%0A%09%09%22text%22%3A%20%7B%0A%20%20%20%20%20%20%20%20%20%20%20%20%22type%22%3A%20%22mrkdwn%22%2C%0A%20%20%20%20%20%20%20%20%20%20%20%20%22text%22%3A%20%22A%20message%20*with%20some%20bold%20text*%20and%20_some%20italicized%20text_.%22%0A%20%20%20%20%20%20%20%20%7D%2C%0A%09%09%22block_id%22%3A%20%22text1%22%0A%09%7D%0A%5D)
+     * ```
+     */
+    data class TextOnlySection(
+        /**
+         * One of the compatible [Element] objects. Be sure to confirm the desired element works with section.
+         */
+        val accessory: Element,
+        /**
+         * The text for the block, in the form of a [CompositionObject.Text] text object.
+         * Minimum length for the text in this field is 1 and maximum length is 3000 characters.
+         */
+        val text: CompositionObject.Text
+    )
+
+    /**
+     * Examples:
+     * - A section block containing text fields ([view this in Block Kit Builder](https://api.slack.com/tools/block-kit-builder?blocks=%5B%0A%09%7B%0A%09%09%22type%22%3A%20%22section%22%2C%0A%09%09%22text%22%3A%20%7B%0A%09%09%09%22text%22%3A%20%22A%20message%20*with%20some%20bold%20text*%20and%20_some%20italicized%20text_.%22%2C%0A%09%09%09%22type%22%3A%20%22mrkdwn%22%0A%09%09%7D%2C%0A%09%09%22fields%22%3A%20%5B%0A%09%09%09%7B%0A%09%09%09%09%22type%22%3A%20%22mrkdwn%22%2C%0A%09%09%09%09%22text%22%3A%20%22*Priority*%22%0A%09%09%09%7D%2C%0A%09%09%09%7B%0A%09%09%09%09%22type%22%3A%20%22mrkdwn%22%2C%0A%09%09%09%09%22text%22%3A%20%22*Type*%22%0A%09%09%09%7D%2C%0A%09%09%09%7B%0A%09%09%09%09%22type%22%3A%20%22plain_text%22%2C%0A%09%09%09%09%22text%22%3A%20%22High%22%0A%09%09%09%7D%2C%0A%09%09%09%7B%0A%09%09%09%09%22type%22%3A%20%22plain_text%22%2C%0A%09%09%09%09%22text%22%3A%20%22String%22%0A%09%09%09%7D%0A%09%09%5D%0A%09%7D%0A%5D)):
+     * ```json
+     * {
+     *   "type": "section",
+     *   "text": {
+     *     "text": "A message *with some bold text* and _some italicized text_.",
+     *     "type": "mrkdwn"
+     *   },
+     *   "fields": [
+     *     {
+     *       "type": "mrkdwn",
+     *       "text": "High"
+     *     },
+     *     {
+     *       "type": "plain_text",
+     *       "emoji": true,
+     *       "text": "String"
+     *     }
+     *   ]
+     * }
+     * ```
+     * - A section block containing a date picker element ([view this in Block Kit Builder](https://api.slack.com/tools/block-kit-builder?blocks=%5B%0A%09%7B%0A%20%20%22type%22%3A%20%22section%22%2C%0A%20%20%22text%22%3A%20%7B%0A%20%20%20%20%22text%22%3A%20%22*Sally*%20has%20requested%20you%20set%20the%20deadline%20for%20the%20Nano%20launch%20project%22%2C%0A%20%20%20%20%22type%22%3A%20%22mrkdwn%22%0A%20%20%7D%2C%0A%20%20%22accessory%22%3A%20%7B%0A%20%20%20%20%22type%22%3A%20%22datepicker%22%0A%20%20%7D%0A%7D%0A%5D)):
+     * ```json
+     * {
+     *   "type": "section",
+     *   "text": {
+     *     "text": "*Sally* has requested you set the deadline for the Nano launch project",
+     *     "type": "mrkdwn"
+     *   },
+     *   "accessory": {
+     *     "type": "datepicker",
+     *     "action_id": "datepicker123",
+     *     "initial_date": "1990-04-28",
+     *     "placeholder": {
+     *       "type": "plain_text",
+     *       "text": "Select a date"
+     *     }
+     *   }
+     * }
+     *
+     * ```
+     */
+    data class FieldsSection(
+        /**
+         * A list of text objects.
+         * Any text objects included with fields will be rendered in a compact format that allows for 2 columns of
+         * side-by-side text.
+         * Maximum number of items is 10.
+         * Maximum length for the text in each item is 2000 characters.
+         * [Click here for an example](https://api.slack.com/tools/block-kit-builder?blocks=%5B%0A%09%7B%0A%09%09%22type%22%3A%20%22section%22%2C%0A%09%09%22text%22%3A%20%7B%0A%09%09%09%22text%22%3A%20%22A%20message%20*with%20some%20bold%20text*%20and%20_some%20italicized%20text_.%22%2C%0A%09%09%09%22type%22%3A%20%22mrkdwn%22%0A%09%09%7D%2C%0A%09%09%22fields%22%3A%20%5B%0A%09%09%09%7B%0A%09%09%09%09%22type%22%3A%20%22mrkdwn%22%2C%0A%09%09%09%09%22text%22%3A%20%22*Priority*%22%0A%09%09%09%7D%2C%0A%09%09%09%7B%0A%09%09%09%09%22type%22%3A%20%22mrkdwn%22%2C%0A%09%09%09%09%22text%22%3A%20%22*Type*%22%0A%09%09%09%7D%2C%0A%09%09%09%7B%0A%09%09%09%09%22type%22%3A%20%22plain_text%22%2C%0A%09%09%09%09%22text%22%3A%20%22High%22%0A%09%09%09%7D%2C%0A%09%09%09%7B%0A%09%09%09%09%22type%22%3A%20%22plain_text%22%2C%0A%09%09%09%09%22text%22%3A%20%22String%22%0A%09%09%09%7D%0A%09%09%5D%0A%09%7D%0A%5D).
+         */
+        val fields: List<CompositionObject.Text>,
+        /**
+         * The text for the block, in the form of a [CompositionObject.Text] text object.
+         * Minimum length for the text in this field is 1 and maximum length is 3000 characters.
+         */
+        val text: CompositionObject.Text?
+    )
+}
