+++ src/main/kotlin/eth/rochsolid/slack/blocks/models/elements/RichText.kt
@@ -1,0 +1,5 @@
+package eth.rochsolid.slack.blocks.models.elements
+
+sealed interface RichText {
+    sealed interface Input
+}
